# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MarkdownTown API",
    "version": "0.1.0",
    "description": "Sections CRUD and auth endpoints"
  },
  "servers": [{ "url": "https://markdown.town" }],
  "paths": {
    "/api/sections": {
      "get": {
        "summary": "List sections",
        "responses": {
          "200": {
            "description": "List of sections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Section" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create section",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SectionInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Section created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Section" }
              }
            }
          }
        }
      }
    },
    "/api/sections/{id}": {
      "put": {
        "summary": "Update section",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SectionInput" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated section",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Section" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete section",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Deleted" }
        }
      }
    },
    "/api/auth/signin": {
      "get": {
        "summary": "Start GitHub OAuth",
        "responses": {
          "302": { "description": "Redirect to provider" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Section": {
        "type": "object",
        "required": ["id", "title", "content", "order", "userId"],
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "order": { "type": "integer" },
          "tags": {
            "type": "array",
            "items": { "type": "string", "maxLength": 32 },
            "maxItems": 8
          },
          "userId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SectionInput": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "content": { "type": "string" },
          "tags": {
            "type": "array",
            "items": { "type": "string", "maxLength": 32 },
            "maxItems": 8
          }
        }
      }
    }
  }
}